Update Version 8
In version 8 of my program, I made a substantial improvement by introducing separate windows for different functionalities, such as printing and deleting details. This approach not only makes the program more advanced but also enhances user convenience. Also, I have added the validated input to the detail function, so only integers are able to show up, as well as a error message popping up if none of the existing receipt number matches with the input. While the idea and code structure were inspired by a friend’s work (with their permission), I used the ideas and modified it to suit my needs(did not copy).