Update Version 9
In version 9, I have created a random receipt number generator that assigns a number between 1 and 100, replacing the "increasing by one" receipt numbering system from previous versions. As a result, the receipt number counter is no longer needed, so I removed it. Additionally, unlike in version 8, I created a new window specifically for appending details, using Toplevel, similar to how the detail and print windows are handled. Also, for the new Append detail window, a Frame is used to locate the buttons and labels to the middle of the window.