Create Version 15
In version 15, I made some debugging adjustments to the delete function. Since I added an extra entry for the last name, the detail list for the delete function also needed to be changed. Additionally, I added an error message that appears when the inputted receipt number does not match any existing hire details.